requirements.txt
image==1.5.33
mouse==0.7.1
MouseInfo==0.1.3
mss==9.0.1
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyscreenshot==3.1
PyScreeze==0.1.30
pytweening==1.0.7

